Rename active deployment path
